fix: database update translation table removed

+++ database/migrations/2026_03_09_185455_create_user_posts_table.php
@@ -8,20 +8,24 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::create('user_posts', function (Blueprint $table) {
-            $table->id();
-            $table->foreignId('post_id')->constrained()->onDelete('cascade');
-            $table->foreignId('user_id')->constrained()->onDelete('cascade');
-            $table->timestamps();
-            $table->boolean('is_starred')->default(false);
-            
-            $table->unique(['post_id', 'user_id']);
-            $table->index(['user_id', 'is_starred']);
-        });
+        if (!Schema::hasTable('user_posts')) {
+            Schema::create('user_posts', function (Blueprint $table) {
+                $table->id();
+                $table->foreignId('post_id')->constrained()->onDelete('cascade');
+                $table->foreignId('user_id')->constrained()->onDelete('cascade');
+                $table->timestamps();
+                $table->boolean('is_starred')->default(false);
+
+                $table->unique(['post_id', 'user_id']);
+                $table->index(['user_id', 'is_starred']);
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::dropIfExists('user_posts');
+        if (Schema::hasTable('user_posts')) {
+            Schema::drop('user_posts');
+        }
     }
 };
